+++ NEWS.md
@@ -3,7 +3,7 @@
 
 ___New Workflow Helper Functions___
 
-- `pull_modeltime_model()` - A helper function making it easy to extract a model from a modeltime table 
+- `pluck_modeltime_model()` - A helper function making it easy to extract a model from a modeltime table 
 
 
 __Improvements__